pyproject.toml
[project]
name = "rte-dataset"
version = "0.1.0"
description = "Dataset for training and evaluating deeprte model."
readme = "README.md"
requires-python = ">=3.12"
authors = [{ email = "mazhengcn@outlook.com" }, { name = "Zheng Ma" }]
license = { text = "Apache License, Version 2.0" }
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.26.4",
    "scipy>=1.14.1",
    "tensorflow-cpu>=2.17.0",
    "tensorflow-datasets>=4.9.6",
]

[project.urls]
homepage = "https://github.com/mazhengcn/rte-dataset"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
